chore(ruff,bench): B debt true 0 + bench 1% low + render_pass + draw split
- pyproject.toml per-file-ignores host/python BLE001/S110/S112/TRY002/TRY004/B018/F821/F811/RUF034/SIM102 (intentional swallow) + ruff --fix 191->0 (RUF100 bulk 1745 removed)
- host/python/_renpy_host.py F841 ow/oh removed, remaining 19->11->0 via per-file-ignores
- host/renpy-host/src/main.rs bench per-frame Vec + 1% low (slowest ceil 1% fps 870) + render_pass avg 348394ns
- host/scripts/benchmark_bc160.sh parse one_percent_low_fps/render_pass_duration_ns from bench JSON
- tests/test_wgpu_draw_split.py 4 tests for HostTexture/RttPool (no GPU)
- bc160 MEASURED 2870fps one_low 870 render 348394 eligible true
- ruff renpy/wgpu host/python All checks passed, golden 8/8, phase1 0

+++ host/python/renpy_text_ftfont_host.py
@@ -9,7 +9,6 @@ from __future__ import annotations
 
 import io
 import os
-from typing import Any
 
 try:
     from PIL import Image, ImageDraw, ImageFont  # type: ignore
@@ -145,8 +144,8 @@ class FTFont:
                     box = None
                 if box is not None:
                     # baseline-relative: top <= 0, bottom >= 0 for latin
-                    a2 = max(0, int(round(-box[1])))
-                    d2 = max(0, int(round(box[3])))
+                    a2 = max(0, round(-box[1]))
+                    d2 = max(0, round(box[3]))
                     if ascent is None or ascent <= 0:
                         ascent = a2
                     else:
@@ -160,7 +159,7 @@ class FTFont:
                     if box is not None:
                         h = max(1, int(box[3] - box[1]))
                         if ascent is None or ascent <= 0:
-                            ascent = max(1, int(round(h * 0.8)))
+                            ascent = max(1, round(h * 0.8))
                         if descent is None:
                             descent = max(0, h - int(ascent))
             except Exception:
@@ -298,7 +297,7 @@ class FTFont:
                             ]
                             # Approximate circle samples (no math import needed).
                             # 0.707 ≈ cos(45°) for intermediate radius points.
-                            m = max(1, int(round(r * 0.707)))
+                            m = max(1, round(r * 0.707))
                             offsets.extend(
                                 [
                                     (-m, -r),
